Optimize greg to jul days rebasing
